requirements.txt
# Core requirements (SQLite mode)
# For Postgres mode: pip install zikra[postgres]
# For Full profile: pip install zikra[full]
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
sqlite-vec>=0.1.1
python-dotenv>=1.0.0
mcp[cli]>=1.0.0
starlette>=0.37.0

# optional: only needed for DB_BACKEND=postgres
# asyncpg>=0.29.0
